Code consolidation and cleanup
