+++ src/app/SystemFolder/SystemResources/File/ClassicyFileBrowserViewIcons.tsx
@@ -2,8 +2,7 @@ import { useDesktop, useDesktopDispatch } from '@/app/SystemFolder/ControlPanels
 import { ClassicyFileSystem } from '@/app/SystemFolder/SystemResources/File/ClassicyFileSystem'
 import ClassicyIcon from '@/app/SystemFolder/SystemResources/Icon/ClassicyIcon'
 import React, { RefObject, useEffect, useState } from 'react'
-import { getIconSize } from '@/app/SystemFolder/SystemResources/Desktop/ClassicyDesktopIconContext'
-import { ClassicyTheme } from '@/app/SystemFolder/ControlPanels/AppearanceManager/ClassicyAppearance'
+import { cleanupIcon, iconImageByType } from '@/app/SystemFolder/SystemResources/File/ClassicyFileBrowserUtils'
 
 type ClassicyFileBrowserViewIconsProps = {
     fs: ClassicyFileSystem
@@ -27,17 +26,6 @@ const ClassicyFileBrowserViewIcons: React.FC<ClassicyFileBrowserViewIconsProps> 
 
     const [items, setItems] = useState([])
 
-    const iconImageByType = (byType: string) => {
-        switch (byType) {
-            case 'directory': {
-                return `${process.env.NEXT_PUBLIC_BASE_PATH || ''}/img/icons/system/folders/directory.png`
-            }
-            default: {
-                return `${process.env.NEXT_PUBLIC_BASE_PATH || ''}/img/icons/system/files/file.png`
-            }
-        }
-    }
-
     const openFileOrFolder = (properties, path: string, filename: string) => {
         switch (properties['_type']) {
             case 'directory': {
@@ -50,34 +38,6 @@ const ClassicyFileBrowserViewIcons: React.FC<ClassicyFileBrowserViewIconsProps> 
                 return () => {}
             }
         }
-    }
-
-    const createGrid = (
-        iconSize: number,
-        iconPadding: number,
-        containerMeasure: [number, number]
-    ): [number, number] => {
-        return [
-            Math.floor(containerMeasure[0] / (iconSize * 2 + iconPadding)),
-            Math.floor(containerMeasure[1] / (iconSize * 2 + iconPadding)),
-        ]
-    }
-
-    const getGridPosition = (i: number, grid: [number, number]): [number, number] => {
-        return [i % grid[0], Math.floor(i / grid[0])]
-    }
-
-    function cleanupIcon(
-        theme: ClassicyTheme,
-        iconIndex: number,
-        iconTotal: number,
-        containerMeasure: [number, number]
-    ): [number, number] {
-        const [iconSize, iconPadding] = getIconSize(theme)
-        let grid = createGrid(iconSize, iconTotal, containerMeasure)
-        const [startX, startY] = getGridPosition(iconIndex, grid)
-
-        return [iconPadding + Math.floor(iconSize * 2 * startX), iconPadding + Math.floor(iconSize * 2 * startY)]
     }
 
     useEffect(() => {
